task5/requirements.txt
google-cloud-bigquery
google-cloud-logging
google-generativeai
vertexai
flask
requests
pytest
pandas